+++ src/pg/pg-store.ts
@@ -1,24 +1,26 @@
 import { BitcoinEvent, Payload } from '@hirosystems/chainhook-client';
 import { Order, OrderBy } from '../api/schemas';
 import { isProdEnv, isTestEnv, normalizedHexString, parseSatPoint } from '../api/util/helpers';
-import { OrdinalSatoshi, SatoshiRarity } from '../api/util/ordinal-satoshi';
+import { OrdinalSatoshi } from '../api/util/ordinal-satoshi';
 import { ENV } from '../env';
-import { getIndexResultCountType } from './helpers';
+import { getInscriptionRecursion } from './helpers';
 import {
   DbFullyLocatedInscriptionResult,
+  DbInscription,
   DbInscriptionContent,
   DbInscriptionCountPerBlock,
   DbInscriptionCountPerBlockFilters,
   DbInscriptionIndexFilters,
   DbInscriptionIndexOrder,
   DbInscriptionIndexPaging,
-  DbInscriptionIndexResultCountType,
   DbInscriptionInsert,
   DbInscriptionLocationChange,
   DbLocation,
   DbLocationInsert,
+  DbLocationPointer,
   DbLocationPointerInsert,
   DbPaginatedResult,
+  INSCRIPTIONS_COLUMNS,
   LOCATIONS_COLUMNS,
 } from './types';
 import {
@@ -30,6 +32,9 @@ import {
 } from '@hirosystems/api-toolkit';
 import * as path from 'path';
 import { Brc20PgStore } from './brc20/brc20-pg-store';
+import { CountsPgStore } from './counts/counts-pg-store';
+import { getIndexResultCountType } from './counts/helpers';
+import * as postgres from 'postgres';
 
 export const MIGRATIONS_DIR = path.join(__dirname, '../../migrations');
 
@@ -37,6 +42,7 @@ type InscriptionIdentifier = { genesis_id: string } | { number: number };
 
 export class PgStore extends BasePgStore {
   readonly brc20: Brc20PgStore;
+  readonly counts: CountsPgStore;
 
   static async connect(opts?: { skipMigrations: boolean }): Promise<PgStore> {
     const pgConfig = {
@@ -64,6 +70,7 @@ export class PgStore extends BasePgStore {
   constructor(sql: PgSqlClient) {
     super(sql);
     this.brc20 = new Brc20PgStore(this);
+    this.counts = new CountsPgStore(this);
   }
 
   /**
@@ -211,10 +218,6 @@ export class PgStore extends BasePgStore {
       // we can respond to the chainhook node with a `200` HTTP code as soon as possible.
       const viewRefresh = Promise.allSettled([
         this.normalizeInscriptionCount({ min_block_height: updatedBlockHeightMin }),
-        this.refreshMaterializedView('inscription_count'),
-        this.refreshMaterializedView('address_counts'),
-        this.refreshMaterializedView('mime_type_counts'),
-        this.refreshMaterializedView('sat_rarity_counts'),
       ]);
       // Only wait for these on tests.
       if (isTestEnv) await viewRefresh;
@@ -224,43 +227,6 @@ export class PgStore extends BasePgStore {
   async getChainTipBlockHeight(): Promise<number> {
     const result = await this.sql<{ block_height: string }[]>`SELECT block_height FROM chain_tip`;
     return parseInt(result[0].block_height);
-  }
-
-  async getChainTipInscriptionCount(): Promise<number> {
-    const result = await this.sql<{ count: number }[]>`
-      SELECT count FROM inscription_count
-    `;
-    return result[0].count;
-  }
-
-  async getMimeTypeInscriptionCount(mimeType?: string[]): Promise<number> {
-    if (!mimeType) return 0;
-    const result = await this.sql<{ count: number }[]>`
-      SELECT COALESCE(SUM(count), 0) AS count
-      FROM mime_type_counts
-      WHERE mime_type IN ${this.sql(mimeType)}
-    `;
-    return result[0].count;
-  }
-
-  async getSatRarityInscriptionCount(satRarity?: SatoshiRarity[]): Promise<number> {
-    if (!satRarity) return 0;
-    const result = await this.sql<{ count: number }[]>`
-      SELECT COALESCE(SUM(count), 0) AS count
-      FROM sat_rarity_counts
-      WHERE sat_rarity IN ${this.sql(satRarity)}
-    `;
-    return result[0].count;
-  }
-
-  async getAddressInscriptionCount(address?: string[]): Promise<number> {
-    if (!address) return 0;
-    const result = await this.sql<{ count: number }[]>`
-      SELECT COALESCE(SUM(count), 0) AS count
-      FROM address_counts
-      WHERE address IN ${this.sql(address)}
-    `;
-    return result[0].count;
   }
 
   async getMaxInscriptionNumber(): Promise<number | undefined> {
@@ -336,22 +302,24 @@ export class PgStore extends BasePgStore {
     sort?: DbInscriptionIndexOrder
   ): Promise<DbPaginatedResult<DbFullyLocatedInscriptionResult>> {
     return await this.sqlTransaction(async sql => {
-      // Do we need a filtered `COUNT(*)`? If so, try to use the pre-calculated counts we have in
-      // materialized views to speed up these queries.
-      const countType = getIndexResultCountType(filters);
-      // `ORDER BY` statement
-      let orderBy = sql`gen.block_height`;
+      const order = sort?.order === Order.asc ? sql`ASC` : sql`DESC`;
+      let orderBy = sql`i.number ${order}`;
       switch (sort?.order_by) {
+        case OrderBy.genesis_block_height:
+          orderBy = sql`gen.block_height ${order}, i.number DESC`;
+          break;
         case OrderBy.ordinal:
-          orderBy = sql`i.sat_ordinal`;
+          orderBy = sql`i.sat_ordinal ${order}`;
           break;
         case OrderBy.rarity:
-          orderBy = sql`ARRAY_POSITION(ARRAY['common','uncommon','rare','epic','legendary','mythic'], i.sat_rarity)`;
+          orderBy = sql`ARRAY_POSITION(ARRAY['common','uncommon','rare','epic','legendary','mythic'], i.sat_rarity) ${order}, i.number DESC`;
           break;
       }
-      // `ORDER` statement
-      const order = sort?.order === Order.asc ? sql`ASC` : sql`DESC`;
-      const results = await sql<({ total: number } & DbFullyLocatedInscriptionResult)[]>`
+      // This function will generate a query to be used for getting results or total counts.
+      const query = (
+        columns: postgres.PendingQuery<postgres.Row[]>,
+        sorting: postgres.PendingQuery<postgres.Row[]>
+      ) => sql`
         WITH gen_locations AS (
           SELECT l.* FROM locations AS l
           INNER JOIN genesis_locations AS g ON l.id = g.location_id
@@ -360,33 +328,7 @@ export class PgStore extends BasePgStore {
           SELECT l.* FROM locations AS l
           INNER JOIN current_locations AS c ON l.id = c.location_id
         )
-        SELECT
-          i.genesis_id,
-          i.number,
-          i.mime_type,
-          i.content_type,
-          i.content_length,
-          i.fee AS genesis_fee,
-          i.curse_type,
-          i.sat_ordinal,
-          i.sat_rarity,
-          i.sat_coinbase_height,
-          gen.block_height AS genesis_block_height,
-          gen.block_hash AS genesis_block_hash,
-          gen.tx_id AS genesis_tx_id,
-          gen.timestamp AS genesis_timestamp,
-          gen.address AS genesis_address,
-          loc.tx_id,
-          loc.address,
-          loc.output,
-          loc.offset,
-          loc.timestamp,
-          loc.value,
-          ${
-            countType === DbInscriptionIndexResultCountType.singleResult
-              ? sql`COUNT(*) OVER() as total`
-              : sql`0 as total`
-          }
+        SELECT ${columns}
         FROM inscriptions AS i
         INNER JOIN cur_locations AS loc ON loc.inscription_id = i.id
         INNER JOIN gen_locations AS gen ON gen.inscription_id = i.id
@@ -454,24 +396,57 @@ export class PgStore extends BasePgStore {
               : sql``
           }
           ${filters?.sat_ordinal ? sql`AND i.sat_ordinal = ${filters.sat_ordinal}` : sql``}
-        ORDER BY ${orderBy} ${order}
-        LIMIT ${page.limit}
-        OFFSET ${page.offset}
+          ${filters?.recursive !== undefined ? sql`AND i.recursive = ${filters.recursive}` : sql``}
+          ${filters?.cursed === true ? sql`AND i.number < 0` : sql``}
+          ${filters?.cursed === false ? sql`AND i.number >= 0` : sql``}
+          ${
+            filters?.genesis_address?.length
+              ? sql`AND gen.address IN ${sql(filters.genesis_address)}`
+              : sql``
+          }
+        ${sorting}
       `;
-      let total = results[0]?.total ?? 0;
-      switch (countType) {
-        case DbInscriptionIndexResultCountType.all:
-          total = await this.getChainTipInscriptionCount();
-          break;
-        case DbInscriptionIndexResultCountType.mimeType:
-          total = await this.getMimeTypeInscriptionCount(filters?.mime_type);
-          break;
-        case DbInscriptionIndexResultCountType.satRarity:
-          total = await this.getSatRarityInscriptionCount(filters?.sat_rarity);
-          break;
-        case DbInscriptionIndexResultCountType.address:
-          total = await this.getAddressInscriptionCount(filters?.address);
-          break;
+      const results = await sql<DbFullyLocatedInscriptionResult[]>`${query(
+        sql`
+          i.genesis_id,
+          i.number,
+          i.mime_type,
+          i.content_type,
+          i.content_length,
+          i.fee AS genesis_fee,
+          i.curse_type,
+          i.sat_ordinal,
+          i.sat_rarity,
+          i.sat_coinbase_height,
+          i.recursive,
+          (
+            SELECT STRING_AGG(ii.genesis_id, ',')
+            FROM inscription_recursions AS ir
+            INNER JOIN inscriptions AS ii ON ii.id = ir.ref_inscription_id
+            WHERE ir.inscription_id = i.id
+          ) AS recursion_refs,
+          gen.block_height AS genesis_block_height,
+          gen.block_hash AS genesis_block_hash,
+          gen.tx_id AS genesis_tx_id,
+          gen.timestamp AS genesis_timestamp,
+          gen.address AS genesis_address,
+          loc.tx_id,
+          loc.address,
+          loc.output,
+          loc.offset,
+          loc.timestamp,
+          loc.value
+        `,
+        sql`ORDER BY ${orderBy} LIMIT ${page.limit} OFFSET ${page.offset}`
+      )}`;
+      // Do we need a filtered `COUNT(*)`? If so, try to use the pre-calculated counts we have in
+      // cached tables to speed up these queries.
+      const countType = getIndexResultCountType(filters);
+      let total = await this.counts.fromResults(countType, filters);
+      if (total === undefined) {
+        // If the count is more complex, attempt it with a separate query.
+        const count = await sql<{ total: number }[]>`${query(sql`COUNT(*) AS total`, sql``)}`;
+        total = count[0].total;
       }
       return {
         total,
@@ -520,7 +495,10 @@ export class PgStore extends BasePgStore {
             FROM locations AS ll
             WHERE
               ll.inscription_id = i.id
-              AND ll.block_height < l.block_height
+              AND (
+                ll.block_height < l.block_height OR
+                (ll.block_height = l.block_height AND ll.tx_index < l.tx_index)
+              )
             ORDER BY ll.block_height DESC
             LIMIT 1
           ) AS from_id,
@@ -547,6 +525,7 @@ export class PgStore extends BasePgStore {
       FROM transfers AS t
       INNER JOIN locations AS lf ON t.from_id = lf.id
       INNER JOIN locations AS lt ON t.to_id = lt.id
+      ORDER BY to_tx_index DESC
     `;
     return {
       total: results[0]?.total ?? 0,
@@ -585,6 +564,29 @@ export class PgStore extends BasePgStore {
     } ${this.sql(viewName)}`;
   }
 
+  private async getInscription(args: { genesis_id: string }): Promise<DbInscription | undefined> {
+    const query = await this.sql<DbInscription[]>`
+      SELECT ${this.sql(INSCRIPTIONS_COLUMNS)}
+      FROM inscriptions
+      WHERE genesis_id = ${args.genesis_id}
+    `;
+    if (query.count === 0) return;
+    return query[0];
+  }
+
+  private async getLocation(args: {
+    genesis_id: string;
+    output: string;
+  }): Promise<DbLocation | undefined> {
+    const query = await this.sql<DbLocation[]>`
+      SELECT ${this.sql(LOCATIONS_COLUMNS)}
+      FROM locations
+      WHERE genesis_id = ${args.genesis_id} AND output = ${args.output}
+    `;
+    if (query.count === 0) return;
+    return query[0];
+  }
+
   private async insertInscription(args: {
     inscription: DbInscriptionInsert;
     location: DbLocationInsert;
@@ -594,8 +596,13 @@ export class PgStore extends BasePgStore {
       const upsert = await sql<{ id: number }[]>`
         SELECT id FROM inscriptions WHERE number = ${args.inscription.number}
       `;
+      const recursion = getInscriptionRecursion(args.inscription.content);
+      const data = {
+        ...args.inscription,
+        recursive: recursion.length > 0,
+      };
       const inscription = await sql<{ id: number }[]>`
-        INSERT INTO inscriptions ${sql(args.inscription)}
+        INSERT INTO inscriptions ${sql(data)}
         ON CONFLICT ON CONSTRAINT inscriptions_number_unique DO UPDATE SET
           genesis_id = EXCLUDED.genesis_id,
           mime_type = EXCLUDED.mime_type,
@@ -611,18 +618,8 @@ export class PgStore extends BasePgStore {
       `;
       inscription_id = inscription[0].id;
       const location = {
+        ...args.location,
         inscription_id,
-        genesis_id: args.location.genesis_id,
-        block_height: args.location.block_height,
-        block_hash: args.location.block_hash,
-        tx_id: args.location.tx_id,
-        tx_index: args.location.tx_index,
-        address: args.location.address,
-        output: args.location.output,
-        offset: args.location.offset,
-        prev_output: args.location.prev_output,
-        prev_offset: args.location.prev_offset,
-        value: args.location.value,
         timestamp: sql`to_timestamp(${args.location.timestamp})`,
       };
       const locationRes = await sql<{ id: number }[]>`
@@ -653,6 +650,8 @@ export class PgStore extends BasePgStore {
         tx_index: args.location.tx_index,
         address: args.location.address,
       });
+      await this.updateInscriptionRecursion({ inscription_id, ref_genesis_ids: recursion });
+      await this.counts.applyInscription({ inscription: args.inscription });
       logger.info(
         `PgStore${upsert.count > 0 ? ' upsert ' : ' '}reveal #${args.inscription.number} (${
           args.location.genesis_id
@@ -678,8 +677,7 @@ export class PgStore extends BasePgStore {
       if (args.location.prev_output) {
         const prev = await sql`
           SELECT id FROM locations
-          WHERE genesis_id = ${args.location.genesis_id}
-            AND prev_output = ${args.location.prev_output}
+          WHERE genesis_id = ${args.location.genesis_id} AND output = ${args.location.prev_output}
         `;
         if (prev.count === 0) {
           logger.warn(
@@ -689,8 +687,7 @@ export class PgStore extends BasePgStore {
       }
       const upsert = await sql`
         SELECT id FROM locations
-        WHERE genesis_id = ${args.location.genesis_id}
-        AND block_height = ${args.location.block_height}
+        WHERE output = ${args.location.output} AND "offset" = ${args.location.offset}
       `;
       const location = {
         inscription_id,
@@ -789,11 +786,15 @@ export class PgStore extends BasePgStore {
     number: number;
     block_height: number;
   }): Promise<void> {
-    // This will cascade into dependent tables.
-    await this.sql`DELETE FROM inscriptions WHERE genesis_id = ${args.genesis_id}`;
-    logger.info(
-      `PgStore rollback reveal #${args.number} (${args.genesis_id}) at block ${args.block_height}`
-    );
+    const inscription = await this.getInscription({ genesis_id: args.genesis_id });
+    if (!inscription) return;
+    await this.sqlWriteTransaction(async sql => {
+      await this.counts.rollBackInscription({ inscription });
+      await sql`DELETE FROM inscriptions WHERE id = ${inscription.id}`;
+      logger.info(
+        `PgStore rollback reveal #${args.number} (${args.genesis_id}) at block ${args.block_height}`
+      );
+    });
   }
 
   private async rollBackLocation(args: {
@@ -801,13 +802,15 @@ export class PgStore extends BasePgStore {
     output: string;
     block_height: number;
   }): Promise<void> {
-    await this.sql`
-      DELETE FROM locations
-      WHERE genesis_id = ${args.genesis_id} AND output = ${args.output}
-    `;
-    logger.info(
-      `PgStore rollback transfer (${args.genesis_id}) on output ${args.output} at block ${args.block_height}`
-    );
+    const location = await this.getLocation({ genesis_id: args.genesis_id, output: args.output });
+    if (!location) return;
+    await this.sqlWriteTransaction(async sql => {
+      await this.recalculateCurrentLocationPointerFromLocationRollBack({ location });
+      await sql`DELETE FROM locations WHERE id = ${location.id}`;
+      logger.info(
+        `PgStore rollback transfer (${args.genesis_id}) on output ${args.output} at block ${args.block_height}`
+      );
+    });
   }
 
   private async updateInscriptionLocationPointers(
@@ -822,9 +825,13 @@ export class PgStore extends BasePgStore {
         tx_index: args.tx_index,
         address: args.address,
       };
-      await sql`
+
+      const genesis = await sql<DbLocationPointer[]>`
+        SELECT * FROM genesis_locations WHERE inscription_id = ${args.inscription_id}
+      `;
+      const genesisRes = await sql`
         INSERT INTO genesis_locations ${sql(pointer)}
-        ON CONFLICT ON CONSTRAINT genesis_locations_inscription_id_unique DO UPDATE SET
+        ON CONFLICT (inscription_id) DO UPDATE SET
           location_id = EXCLUDED.location_id,
           block_height = EXCLUDED.block_height,
           tx_index = EXCLUDED.tx_index,
@@ -834,9 +841,17 @@ export class PgStore extends BasePgStore {
           (EXCLUDED.block_height = genesis_locations.block_height AND
             EXCLUDED.tx_index < genesis_locations.tx_index)
       `;
-      await sql`
+      // Affect genesis counts only if we have an update.
+      if (genesisRes.count > 0) {
+        await this.counts.applyGenesisLocation({ old: genesis[0], new: pointer });
+      }
+
+      const current = await sql<DbLocationPointer[]>`
+        SELECT * FROM current_locations WHERE inscription_id = ${args.inscription_id}
+      `;
+      const currentRes = await sql`
         INSERT INTO current_locations ${sql(pointer)}
-        ON CONFLICT ON CONSTRAINT current_locations_inscription_id_unique DO UPDATE SET
+        ON CONFLICT (inscription_id) DO UPDATE SET
           location_id = EXCLUDED.location_id,
           block_height = EXCLUDED.block_height,
           tx_index = EXCLUDED.tx_index,
@@ -846,6 +861,11 @@ export class PgStore extends BasePgStore {
           (EXCLUDED.block_height = current_locations.block_height AND
             EXCLUDED.tx_index > current_locations.tx_index)
       `;
+      // Affect current location counts only if we have an update.
+      if (currentRes.count > 0) {
+        await this.counts.applyCurrentLocation({ old: current[0], new: pointer });
+      }
+
       // Backfill orphan locations for this inscription, if any.
       await sql`
         UPDATE locations
@@ -856,6 +876,58 @@ export class PgStore extends BasePgStore {
       await sql`
         UPDATE inscriptions SET updated_at = NOW() WHERE genesis_id = ${args.genesis_id}
       `;
+    });
+  }
+
+  private async recalculateCurrentLocationPointerFromLocationRollBack(args: {
+    location: DbLocation;
+  }): Promise<void> {
+    await this.sqlWriteTransaction(async sql => {
+      // Is the location we're rolling back *the* current location?
+      const current = await sql<DbLocationPointer[]>`
+        SELECT * FROM current_locations WHERE location_id = ${args.location.id}
+      `;
+      if (current.count > 0) {
+        const update = await sql<DbLocationPointer[]>`
+          WITH prev AS (
+            SELECT id, block_height, tx_index, address
+            FROM locations
+            WHERE inscription_id = ${args.location.inscription_id} AND id <> ${args.location.id}
+            ORDER BY block_height DESC, tx_index DESC
+            LIMIT 1
+          )
+          UPDATE current_locations AS c SET
+            location_id = prev.id,
+            block_height = prev.block_height,
+            tx_index = prev.tx_index,
+            address = prev.address
+          FROM prev
+          WHERE c.inscription_id = ${args.location.inscription_id}
+          RETURNING *
+        `;
+        await this.counts.rollBackCurrentLocation({ curr: current[0], prev: update[0] });
+      }
+    });
+  }
+
+  private async updateInscriptionRecursion(args: {
+    inscription_id: number;
+    ref_genesis_ids: string[];
+  }): Promise<void> {
+    await this.sqlWriteTransaction(async sql => {
+      const validated = await sql<{ id: string }[]>`
+        SELECT id FROM inscriptions WHERE genesis_id IN ${this.sql(args.ref_genesis_ids)}
+      `;
+      if (validated.count > 0) {
+        const values = validated.map(i => ({
+          inscription_id: args.inscription_id,
+          ref_inscription_id: i.id,
+        }));
+        await this.sql`
+          INSERT INTO inscription_recursions ${sql(values)}
+          ON CONFLICT ON CONSTRAINT inscriptions_inscription_id_ref_inscription_id_unique DO NOTHING
+        `;
+      }
     });
   }
 }